Moved web native implemenration into src, added more API support
