+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # Release Notes
 
-- Added web support for isCameraPresent, isLocationEnabled, isAirplaneMode and battery/charging-related APIs.
+- Added web support for isCameraPresent, isLocationEnabled, isAirplaneMode, getBaseOs, getFreeDiskStorage, getMaxMemory, getTotalDiskCapacity, getTotalMemory, getUsedMemory and battery/charging-related APIs.
 
 ## 5.2.0
